refactor: split MessageListView into 4 focused files
Extract 3 concerns from the 1305-line MessageListView.swift:

- MessageCellView.swift (246 lines): Per-message cell with Equatable
  diffing, previously private — now internal for cross-file access
- MessageListDerivedState.swift (52 lines): CachedMessageLayoutMetadata
  and MessageListDerivedState data types
- MessageListEnvironment.swift (87 lines): SuppressAutoScrollKey,
  ScrollTrackingState, PrecomputedCacheKey

MessageListView.swift drops from 1305 to 922 lines. Pure mechanical
refactor — no behavior changes.
