+++ clients/macos/vellum-assistant/Features/Chat/MessageListView.swift
@@ -7,91 +7,6 @@ import VellumAssistantShared
 private let log = Logger(subsystem: "com.vellum.vellum-assistant", category: "MessageListView")
 private let stallLog = OSLog(subsystem: "com.vellum.assistant", category: "LayoutStall")
 
-// MARK: - Scroll Suppression Environment
-
-/// Environment key that child views (e.g. AssistantProgressView) call to
-/// temporarily suppress auto-scroll-to-bottom during content expansion.
-private struct SuppressAutoScrollKey: EnvironmentKey {
-    static let defaultValue: (() -> Void)? = nil
-}
-
-extension EnvironmentValues {
-    var suppressAutoScroll: (() -> Void)? {
-        get { self[SuppressAutoScrollKey.self] }
-        set { self[SuppressAutoScrollKey.self] = newValue }
-    }
-}
-
-/// Holds scroll-related tracking values that must persist across body
-/// evaluations but must NOT trigger SwiftUI re-renders when updated.
-/// These values serve as dead-zone guards and smoothing state — they
-/// are never read during body evaluation for rendering purposes.
-/// Pattern mirrors non-reactive properties on the coordinator (not @Published).
-@MainActor final class ScrollTrackingState {
-
-    // MARK: - Layout Metadata Cache
-
-    /// Cache key for the last computed `CachedMessageLayoutMetadata`.
-    var cachedLayoutKey: PrecomputedCacheKey?
-    /// Cached structural metadata, returned on cache hit.
-    var cachedLayoutMetadata: CachedMessageLayoutMetadata?
-
-    // MARK: - Version Counter (O(1) fingerprint replacement)
-
-    /// Monotonically increasing counter that replaces the O(n) per-body-eval
-    /// `computeMessageFingerprint()` hash. Incremented when any of the following
-    /// triggers fire:
-    /// - raw `messages.count` changes (new message appended, deletion, or
-    ///   pagination load — also catches paginated-window shifts)
-    /// - visible message count changes (hidden/subagent visibility transitions)
-    /// - `isSending` or `isThinking` transitions (activity state change)
-    /// - `visibleMessages.last?.isStreaming` transitions (end of streaming)
-    /// - A tool call's `isComplete` transitions (observable via messages array
-    ///   identity change in SwiftUI)
-    ///
-    /// Over-invalidation is safe (triggers a recompute); under-invalidation is not.
-    var messageListVersion: Int = 0
-
-    /// Cached raw (unfiltered) message count. Detects new arrivals,
-    /// deletions, and pagination loads — including cases where the
-    /// paginated visible window shifts at a fixed size.
-    var lastKnownRawMessageCount: Int = 0
-    /// Cached visible (paginated) message count. Detects changes in
-    /// message visibility (hidden/subagent transitions) that don't
-    /// alter the raw count.
-    var lastKnownVisibleMessageCount: Int = 0
-    /// Cached streaming state of the last visible message.
-    var lastKnownLastMessageStreaming: Bool = false
-    /// Cached count of incomplete tool calls across visible messages.
-    var lastKnownIncompleteToolCallCount: Int = 0
-
-    // MARK: - Scroll Geometry (non-reactive, updated every scroll tick)
-
-    /// Content height from scroll geometry, used to guard against false
-    /// detaches on short conversations that can't scroll.
-    var scrollContentHeight: CGFloat = 0
-    /// Container (viewport) height from scroll geometry, used alongside
-    /// scrollContentHeight to determine if content is scrollable.
-    var scrollContainerHeight: CGFloat = 0
-    /// Last content offset Y observed by onScrollGeometryChange, used to
-    /// determine scroll direction (increasing offset = scrolling toward older content).
-    var lastScrollContentOffsetY: CGFloat = 0
-}
-
-/// Lightweight key that captures all inputs to `precomputedState`.
-/// All fields are O(1) to compare. The `messageListVersion` counter
-/// replaces the former O(n) hash-based fingerprint — it is incremented
-/// by `onChange` handlers when structural or content changes occur.
-struct PrecomputedCacheKey: Equatable {
-    let messageListVersion: Int
-    let isSending: Bool
-    let isThinking: Bool
-    let isCompacting: Bool
-    let assistantStatusText: String?
-    let activeSubagentFingerprint: Int
-    let displayedMessageCount: Int
-}
-
 /// Lightweight snapshot of scroll geometry for the onScrollGeometryChange
 /// handler. Captures values needed for scroll-direction detection,
 /// scrollable-content detection, and viewport height tracking.
@@ -1004,302 +919,4 @@ struct MessageListView: View {
                 }
             }
     }
-}
-
-// MARK: - MessageCellView
-
-/// Per-message cell extracted from the ForEach body so SwiftUI has a typed
-/// struct boundary for diffing: when all `let` inputs are equal, SwiftUI can
-/// skip re-evaluating the body during LazySubviewPlacements.updateValue.
-private struct MessageCellView: View, Equatable {
-    static func == (lhs: MessageCellView, rhs: MessageCellView) -> Bool {
-        lhs.message == rhs.message
-            && lhs.index == rhs.index
-            && lhs.showTimestamp == rhs.showTimestamp
-            && lhs.nextDecidedConfirmation == rhs.nextDecidedConfirmation
-            && lhs.isConfirmationRenderedInline == rhs.isConfirmationRenderedInline
-            && lhs.hasPrecedingAssistant == rhs.hasPrecedingAssistant
-            && lhs.hasUserMessage == rhs.hasUserMessage
-            && lhs.activePendingRequestId == rhs.activePendingRequestId
-            && lhs.latestAssistantId == rhs.latestAssistantId
-            && lhs.anchoredThinkingIndex == rhs.anchoredThinkingIndex
-            && lhs.subagentsByParent[lhs.message.id] == rhs.subagentsByParent[rhs.message.id]
-            && lhs.canInlineProcessing == rhs.canInlineProcessing
-            && lhs.shouldShowThinkingIndicator == rhs.shouldShowThinkingIndicator
-            && lhs.assistantStatusText == rhs.assistantStatusText
-            && lhs.dismissedDocumentSurfaceIds == rhs.dismissedDocumentSurfaceIds
-            && lhs.activeSurfaceId == rhs.activeSurfaceId
-            && lhs.isHighlighted == rhs.isHighlighted
-            && lhs.selectedModel == rhs.selectedModel
-            && lhs.configuredProviders == rhs.configuredProviders
-            && (lhs.providerCatalogHash != rhs.providerCatalogHash ? false
-                : lhs.providerCatalog.count == rhs.providerCatalog.count
-                  && zip(lhs.providerCatalog, rhs.providerCatalog).allSatisfy({ $0.id == $1.id && $0.displayName == $1.displayName && $0.models.count == $1.models.count && zip($0.models, $1.models).allSatisfy({ $0.id == $1.id && $0.displayName == $1.displayName }) }))
-            && lhs.isTTSEnabled == rhs.isTTSEnabled
-    }
-
-    let message: ChatMessage
-    let index: Int
-    let showTimestamp: Bool
-    let nextDecidedConfirmation: ToolConfirmationData?
-    let isConfirmationRenderedInline: Bool
-    let hasPrecedingAssistant: Bool
-    let hasUserMessage: Bool
-    let activePendingRequestId: String?
-    let latestAssistantId: UUID?
-    let anchoredThinkingIndex: Int?
-    let subagentsByParent: [UUID: [SubagentInfo]]
-    let canInlineProcessing: Bool
-    let shouldShowThinkingIndicator: Bool
-    let assistantStatusText: String?
-    let dismissedDocumentSurfaceIds: Set<String>
-    let activeSurfaceId: String?
-    let isHighlighted: Bool
-    let mediaEmbedSettings: MediaEmbedResolverSettings?
-    let onConfirmationAllow: (String) -> Void
-    let onConfirmationDeny: (String) -> Void
-    let onAlwaysAllow: (String, String, String, String) -> Void
-    var onTemporaryAllow: ((String, String) -> Void)?
-    var onGuardianAction: ((String, String) -> Void)?
-    let onSurfaceAction: (String, String, [String: AnyCodable]?) -> Void
-    let onDismissDocumentWidget: ((String) -> Void)?
-    var onForkFromMessage: ((String) -> Void)?
-    var showInspectButton: Bool = false
-    var isTTSEnabled: Bool = false
-    var onInspectMessage: ((String?) -> Void)?
-    var onRehydrateMessage: ((UUID) -> Void)?
-    var onSurfaceRefetch: ((String, String) -> Void)?
-    var onRetryFailedMessage: ((UUID) -> Void)?
-    var onRetryConversationError: ((UUID) -> Void)?
-    var onAbortSubagent: ((String) -> Void)?
-    var onSubagentTap: ((String) -> Void)?
-    var subagentDetailStore: SubagentDetailStore
-    let selectedModel: String
-    let configuredProviders: Set<String>
-    let providerCatalog: [ProviderCatalogEntry]
-    let providerCatalogHash: Int
-
-    static func hashCatalog(_ catalog: [ProviderCatalogEntry]) -> Int {
-        var hasher = Hasher()
-        for entry in catalog {
-            hasher.combine(entry.id)
-            hasher.combine(entry.displayName)
-            for model in entry.models {
-                hasher.combine(model.id)
-                hasher.combine(model.displayName)
-            }
-        }
-        return hasher.finalize()
-    }
-
-    @AppStorage("hasEverSentMessage") private var hasEverSentMessage: Bool = false
-
-    private func modelListView(for msg: ChatMessage) -> some View {
-        ModelListBubble(currentModel: selectedModel, configuredProviders: configuredProviders, providerCatalog: providerCatalog)
-    }
-
-    private func commandListFallbackMessage(for message: ChatMessage) -> ChatMessage {
-        var fallbackMessage = message
-        fallbackMessage.commandList = nil
-        return fallbackMessage
-    }
-
-    @ViewBuilder
-    private func commandListView(for message: ChatMessage) -> some View {
-        if let commandEntries = CommandListBubble.parsedEntries(from: message.text) {
-            CommandListBubble(commands: commandEntries)
-        } else {
-            ChatBubble(
-                message: commandListFallbackMessage(for: message),
-                decidedConfirmation: nil,
-                onSurfaceAction: onSurfaceAction,
-                onDismissDocumentWidget: { surfaceId in
-                    onDismissDocumentWidget?(surfaceId)
-                },
-                dismissedDocumentSurfaceIds: dismissedDocumentSurfaceIds,
-                onForkFromMessage: onForkFromMessage,
-                showInspectButton: showInspectButton,
-                isTTSEnabled: isTTSEnabled,
-                onInspectMessage: onInspectMessage,
-                onSurfaceRefetch: onSurfaceRefetch,
-                onRehydrate: (message.wasTruncated || message.isContentStripped) ? { onRehydrateMessage?(message.id) } : nil,
-                mediaEmbedSettings: mediaEmbedSettings,
-                onConfirmationAllow: onConfirmationAllow,
-                onConfirmationDeny: onConfirmationDeny,
-                onAlwaysAllow: onAlwaysAllow,
-                onTemporaryAllow: onTemporaryAllow,
-                activeConfirmationRequestId: activePendingRequestId,
-                onRetryFailedMessage: onRetryFailedMessage,
-                onRetryConversationError: message.isError ? { onRetryConversationError?(message.id) } : nil,
-                isLatestAssistantMessage: message.role == .assistant && message.id == latestAssistantId,
-                isProcessingAfterTools: canInlineProcessing && message.id == latestAssistantId,
-                processingStatusText: canInlineProcessing && message.id == latestAssistantId ? assistantStatusText : nil,
-                activeSurfaceId: activeSurfaceId,
-                hideInlineAvatar: shouldShowThinkingIndicator && anchoredThinkingIndex == nil
-            )
-        }
-    }
-
-    @ViewBuilder
-    private func thinkingIndicatorRow() -> some View {
-        RunningIndicator(
-            label: !hasEverSentMessage && hasUserMessage
-                ? "Waking up..."
-                : assistantStatusText ?? "Thinking",
-            showIcon: false
-        )
-        .frame(maxWidth: VSpacing.chatBubbleMaxWidth, alignment: .leading)
-        .id("thinking-indicator")
-    }
-
-    var body: some View {
-        if showTimestamp {
-            TimestampDivider(date: message.timestamp)
-        }
-
-        if let confirmation = message.confirmation {
-            if confirmation.state == .pending {
-                if !isConfirmationRenderedInline {
-                    ToolConfirmationBubble(
-                        confirmation: confirmation,
-                        isKeyboardActive: confirmation.requestId == activePendingRequestId,
-                        onAllow: { onConfirmationAllow(confirmation.requestId) },
-                        onDeny: { onConfirmationDeny(confirmation.requestId) },
-                        onAlwaysAllow: onAlwaysAllow,
-                        onTemporaryAllow: onTemporaryAllow
-                    )
-                    .id(message.id)
-                }
-            } else {
-                if !hasPrecedingAssistant {
-                    ToolConfirmationBubble(
-                        confirmation: confirmation,
-                        onAllow: { onConfirmationAllow(confirmation.requestId) },
-                        onDeny: { onConfirmationDeny(confirmation.requestId) },
-                        onAlwaysAllow: onAlwaysAllow,
-                        onTemporaryAllow: onTemporaryAllow
-                    )
-                    .id(message.id)
-                }
-            }
-        } else if message.modelList != nil {
-            modelListView(for: message)
-                .id(message.id)
-        } else if message.commandList != nil {
-            commandListView(for: message)
-                .id(message.id)
-        } else if let guardianDecision = message.guardianDecision {
-            GuardianDecisionBubble(
-                decision: guardianDecision,
-                onAction: { requestId, action in
-                    onGuardianAction?(requestId, action)
-                }
-            )
-            .id(message.id)
-        } else {
-            ChatBubble(
-                message: message,
-                decidedConfirmation: nextDecidedConfirmation,
-                onSurfaceAction: onSurfaceAction,
-                onDismissDocumentWidget: { surfaceId in
-                    onDismissDocumentWidget?(surfaceId)
-                },
-                dismissedDocumentSurfaceIds: dismissedDocumentSurfaceIds,
-                onForkFromMessage: onForkFromMessage,
-                showInspectButton: showInspectButton,
-                isTTSEnabled: isTTSEnabled,
-                onInspectMessage: onInspectMessage,
-                onSurfaceRefetch: onSurfaceRefetch,
-                onRehydrate: (message.wasTruncated || message.isContentStripped) ? { onRehydrateMessage?(message.id) } : nil,
-                mediaEmbedSettings: mediaEmbedSettings,
-                onConfirmationAllow: onConfirmationAllow,
-                onConfirmationDeny: onConfirmationDeny,
-                onAlwaysAllow: onAlwaysAllow,
-                onTemporaryAllow: onTemporaryAllow,
-                activeConfirmationRequestId: activePendingRequestId,
-                onRetryFailedMessage: onRetryFailedMessage,
-                onRetryConversationError: message.isError ? { onRetryConversationError?(message.id) } : nil,
-                isLatestAssistantMessage: message.role == .assistant && message.id == latestAssistantId,
-                isProcessingAfterTools: canInlineProcessing && message.id == latestAssistantId,
-                processingStatusText: canInlineProcessing && message.id == latestAssistantId ? assistantStatusText : nil,
-                activeSurfaceId: activeSurfaceId,
-                hideInlineAvatar: shouldShowThinkingIndicator && anchoredThinkingIndex == nil
-            )
-            .background(
-                RoundedRectangle(cornerRadius: VRadius.md)
-                    .fill(VColor.primaryBase.opacity(isHighlighted ? 0.15 : 0))
-                    .padding(.horizontal, -VSpacing.sm)
-                    .padding(.vertical, -VSpacing.xs)
-            )
-            .animation(VAnimation.slow, value: isHighlighted)
-            .id(message.id)
-        }
-
-        ForEach(subagentsByParent[message.id] ?? []) { subagent in
-            SubagentEventsReader(
-                store: subagentDetailStore,
-                subagent: subagent,
-                onAbort: { onAbortSubagent?(subagent.id) },
-                onTap: { onSubagentTap?(subagent.id) }
-            )
-                .frame(maxWidth: VSpacing.chatBubbleMaxWidth, alignment: .leading)
-                .id("subagent-\(subagent.id)")
-        }
-
-        if shouldShowThinkingIndicator && anchoredThinkingIndex == index {
-            thinkingIndicatorRow()
-        }
-    }
-}
-
-
-
-// MARK: - Cached Message Layout Metadata
-
-/// Structural metadata cached behind a version-counter key on
-/// `ScrollTrackingState`. Contains only fields derived from message IDs,
-/// roles, timestamps, and subagent identity — never mutable content like
-/// text segments or confirmation states. Cache invalidation is gated by
-/// `refreshMessageListVersionIfNeeded()` which tracks structural changes.
-struct CachedMessageLayoutMetadata {
-    let displayMessageIds: [UUID]
-    let messageIndexById: [UUID: Int]
-    let showTimestamp: Set<UUID>
-    let hasPrecedingAssistantByIndex: Set<Int>
-    let hasUserMessage: Bool
-    let latestAssistantId: UUID?
-    let subagentsByParent: [UUID: [SubagentInfo]]
-    let orphanSubagents: [SubagentInfo]
-    let effectiveStatusText: String?
-}
-
-// MARK: - Message List Derived State
-
-/// All derived values needed by the message list body. Combines cached
-/// structural metadata (from `CachedMessageLayoutMetadata`) with live
-/// content-derived state computed fresh each body evaluation. Content
-/// fields (message data, confirmation placement, thinking indicators)
-/// are always live so SwiftUI's `.equatable()` diffing sees every mutation.
-struct MessageListDerivedState {
-    // --- Cached structural metadata (from CachedMessageLayoutMetadata) ---
-    let displayMessageIds: [UUID]
-    let messageIndexById: [UUID: Int]
-    let showTimestamp: Set<UUID>
-    let hasPrecedingAssistantByIndex: Set<Int>
-    let hasUserMessage: Bool
-    let latestAssistantId: UUID?
-    let subagentsByParent: [UUID: [SubagentInfo]]
-    let orphanSubagents: [SubagentInfo]
-    let effectiveStatusText: String?
-
-    // --- Live content-derived state (always fresh) ---
-    let displayMessageById: [UUID: ChatMessage]
-    let activePendingRequestId: String?
-    let nextDecidedConfirmationByIndex: [Int: ToolConfirmationData]
-    let isConfirmationRenderedInlineByIndex: Set<Int>
-    let anchoredThinkingIndex: Int?
-    let hasActiveToolCall: Bool
-    let canInlineProcessing: Bool
-    let shouldShowThinkingIndicator: Bool
-    let hasMessages: Bool
 }
